Code review feedback
